refactor: consolidate duplicate error note components and fix TypeScript errors
- Move shared error note components from app/error-note/_components/ to app/_components/error-note/
- Fix TypeScript errors by updating FilterParams interface to accept string | null
- Update import paths across all affected files to use new shared location
- Remove empty old components directory
