+++ app/_components/error-note/HelpContentDisplay.tsx
