+++ app/_components/error-note/VideoSection.tsx
