+++ app/_components/error-note/createMyPageConfig.ts
